% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_presence.R
\name{check_file_presence}
\alias{check_file_presence}
\title{Detect the presence of a file within folders}
\usage{
check_file_presence(x, path = getwd())
}
\arguments{
\item{x}{String with the full or partial file name}

\item{path}{Defaults to current directory, can be set to any directory}
}
\value{
TRUE if file is present, FALSE otherwise
}
\description{
Detect the presence of a file within folders
}
\examples{
x <-  "file.csv"
check_file_presence(x)
}
