% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_install_packages.R
\name{check_install_packages}
\alias{check_install_packages}
\title{check_install_packages
check if each package in vector is installed, if not, installs it}
\usage{
check_install_packages(x, y = "https://vps.fmvz.usp.br/CRAN/")
}
\arguments{
\item{x}{vector of package names to be checked an installed}

\item{y}{cran mirror}
}
\value{
install packages if not detected
}
\description{
check_install_packages
check if each package in vector is installed, if not, installs it
}
\examples{

check_install_packages(c("lubridate","devtools"))
}
