% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_backbone_files.R
\name{create_backbone_files}
\alias{create_backbone_files}
\title{Create backbone file system}
\usage{
create_backbone_files()
}
\value{
5 folders, in the current directory
}
\description{
This function creates a template of folders which I am used to,
when starting new projects
}
\examples{
create_backbone_files()
}
