% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List packages
List used packages in the r files of the directory}
\usage{
list_packages()
}
\value{
Vector with package names
}
\description{
List packages
List used packages in the r files of the directory
}
\examples{
No examples
}
