% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{r}
\alias{r}
\alias{qmd}
\alias{rmd}
\title{Create files.}
\usage{
r(name, prefix_by = c("count", "date"))

qmd(name, prefix_by = c("count", "date"))

rmd(name, prefix_by = c("count", "date"))
}
\arguments{
\item{name}{A character string; name of the script.}

\item{prefix_by}{Which prefix to use.}
}
\description{
\code{r()} creates an R script.

\code{rmd()} creates an Rmarkdown document.

\code{qmd()} creates a Quarto document.
}
